+++ rtengine/simpleprocess.cc
@@ -145,7 +145,7 @@ IImage16* processImage (ProcessingJob* pjob, int& errorCode, ProgressListener* p
 
     float* curve1 = new float [65536];
     float* curve2 = new float [65536];
-	int* curve = new int [65536];
+	float* curve = new float [65536];
 	
     CurveFactory::complexCurve (br, bl/65535.0, params.toneCurve.hlcompr, params.toneCurve.shcompr, params.toneCurve.brightness, params.toneCurve.contrast, imgsrc->getDefGain(), imgsrc->getGamma(), true, params.toneCurve.curve, hist16, curve1, curve2, curve, NULL);
